requirements.txt
Flask~=2.2.2
SQLAlchemy~=1.4.41
Flask-Security-Too~=5.0.1
bcrypt~=4.0.0